requirements.txt
Flask==3.0.3
requests==2.32.3
streamlit==1.39.0
python-dotenv==1.0.1